requirements.txt
requests
unittest
mysql-connector-python